% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndlsearch.R
\name{rndlsearch}
\alias{rndlsearch}
\title{Query the Search API of National Diet Library.}
\usage{
rndlsearch(
  limit = NULL,
  title = NULL,
  description = NULL,
  creator = NULL,
  publisher = NULL,
  from = NULL,
  until = NULL,
  ...
)
}
\arguments{
\item{limit}{Number of results. It will be sent as \code{cnt}.}

\item{title}{Title.}

\item{description}{Description.}

\item{creator}{Creator.}

\item{publisher}{Publisher.}

\item{from}{Results after \code{from} specified in \code{yyyy}.}

\item{until}{Results before \code{until} specified in \code{yyyy}.}

\item{...}{Extra parameters. First parameters are preferred if overridden.}
}
\value{
Results. Returns as named \code{list}.
}
\description{
Query the Search API of National Diet Library.
}
